updated server.js to app.js. Added both Steam API routes and they are both sending data back to the FE.
