--- BackEnd/server.js
@@ -1,11 +1,0 @@
-const express = require('express');
-const app = express();
-const port = 3000;
-
-app.get('/', (req, res) => {
-    res.send('hello world!')
-});
-
-app.listen(port, () => {
-    console.log(`listening on port ${port}`)
-})
